closes #3: shiny server Added Shiny Server setup instructions and related changes for prod and dev environment
The updates include the addition of setup instructions, Dockerfile, configuration files, and the start scripts for the Shiny Server in the README.md. This was done to facilitate the deployment and running of Shiny applications in Docker containers. Additionally, changes were made for the user permissions in deployment scripts for security enhancement.

The new Dockerfile provides an image with pre-installed R and required packages for running Shiny applications. The newly added start scripts allow starting the Shiny server in a Docker container with proper configurations and user permissions. The modifications in Caddyfile make the Shiny applications accessible via a reverse proxy.

The Shiny Server setup allows for a structured and consistent way of deploying Shiny applications, speeding up the process and reducing the possibility of errors.