closes #6: bugfix shiny and caddy
Deleted the shiny index app as it was redundant due to Shiny Server's own application discovery feature.
Introduced a Docker network for proper inter-container communication, and modified both the Caddyfile and the start_caddy.sh script to utilize this network correctly.
Fixed the Dockerfile for the Shiny image to set permissions for user apprunner.
Included the shiny_prod and shiny_test scripts to start with the new Docker network and log directory and updated the content of the existing Shiny app's app.R file.
 README.md documentation was also updated to reflect these changes.